+++ src/pages/join-group/helpers.ts
